Review upgrade guide
Incorporates a revised draft by @wolfgangmm and edits by @duncdrum. I added an overview and key concepts section. I removed the historical information on binary compatibility, since this information should be stored in release notes, but this could be restored if it’s important to retain this information in the documentation app itself. (But if so I’d suggest creating a dedicated section for release notes.)